Deduce miptable names from the sqlite3 database
This tries to use the names in the database (and indirectly from the
XML version of the DREQ) to drive the creation and copying scripts.

The database doesn't include a miptable table, so this pulls them from
the mipTable columns of the CMORvar table, which should be the same
thing.  It does mean that any miptables which don't contain any
CMORvars won't get processed, but this is probably fine.